requirements.txt
scapy
python-nmap
rich
